requirements.txt
streamlit
pandas
nbformat
google-api-python-client
google-auth
google-auth-oauthlib
google-generativeai
